+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to
 
 - Added `withTokenBucket` helper function, which executes a callback with an appropriate
   delay according to a provided `HierarchicalTokenBucket`.
+- Added an option to create a child with no options (ie. `child()`) and inherit the 
+  maximumCapacity and refillRate from the parent.
 
 ## 0.1.0 - 2022-05-05